==> Code/1.Concatenate the data.ipynb <==
import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns 

train_1 = pd.read_csv(r"C:\Users\SALAH MAHMOUD\Downloads\ieee-mansb-victoris-2\IEEE_Victoris2_Filtration_train_data\Train_Batch_1.csv")
train_2 = pd.read_csv(r"C:\Users\SALAH MAHMOUD\Downloads\ieee-mansb-victoris-2\IEEE_Victoris2_Filtration_train_data\Train_Batch_2.csv")
train_3 = pd.read_csv(r"C:\Users\SALAH MAHMOUD\Downloads\ieee-mansb-victoris-2\IEEE_Victoris2_Filtration_train_data\Train_Batch_3.csv")

train_1.head(2)

train_2.head(2)

train_3.head(2)

# Prepare data for concat

# The three columns in the dataset have equal values, except for their names.
train_1.rename(columns = {'Min. Yearly Income':'Yearly Income'}, inplace = True)
train_2.rename(columns = {'Min. Person Yearly Income':'Yearly Income'}, inplace = True)

# The column 'Unnamed: 0' is not an important feature.
train_frames = [train_1, train_2, train_3]

# Loop through each DataFrame
for num, train_df in enumerate(train_frames, start=1):
    train_df.drop('Unnamed: 0',axis=1,inplace = True)

# Create a new column named 'Package Weight' for train_2."
train_2['Package Weight'] = train_2['Gross Weight']-train_2['Net Weight']

# Retrieve the 'Package Weight' column from the 'Weights%20Data' dataset.
split_weight_3 = train_3['Weights%20Data']
Package_Weight_3 = [float(i.split(',')[2].split(':')[1].strip('}')) for i in split_weight_3]
train_3['Package Weight'] = Package_Weight_3
train_3.drop(columns='Weights%20Data', inplace=True)

#Retrieve the columns 'Gross Weight', 'Net Weight', and 'Package Weight' from the 'Product Weights Data' dataset in kilograms (KG).

split_weight_1 = train_1['Product Weights Data in (KG)']

Gross_Weight_1 = [float(i.split(',')[0].split(':')[1]) for i in split_weight_1]
Net_Weight_1 = [float(i.split(',')[1].split(':')[1]) for i in split_weight_1]
Package_Weight_1 = [float(i.split(',')[2].split(':')[1].strip('}')) for i in split_weight_1]

train_1['Gross Weight'] = Gross_Weight_1
train_1['Net Weight'] = Net_Weight_1
train_1['Package Weight'] = Package_Weight_1
train_1.drop(columns='Product Weights Data in (KG)', inplace=True)

# Rename columns with %20 replaced by space
for column in train_3.columns:
    new_column_name = column.replace('%20', ' ')
    train_3.rename(columns={column: new_column_name}, inplace=True)

concatenation = pd.concat([train_1, train_2, train_3], ignore_index=True)
df = concatenation.copy()

df.head()
