==> Code/3.Data_Exploration.ipynb <==
import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns 

df = pd.read_csv(r'Preprocessing.csv')

df.head()

df.info()

df.describe()

for col in df.select_dtypes(include=['object']).columns:
    df[col] = df[col].astype('category')

df.isnull().sum().sort_values()

# This code generates a correlation matrix using the Pearson correlation coefficient from a pandas DataFrame "df"
corr_matrix = df.corr(method = "pearson" )

corr_matrix.style.background_gradient(cmap='coolwarm')

#as we see
# high corr (Store Sales,Store Cost),(Store Area,Grocery Area),
#(Frozen Area,Meat Area),(Gross Weight,Net Weight)(Bar For Salad,Ready Food) so we will delete one of thim

df = df.drop(['Frozen Area','Net Weight','Bar For Salad','Store Cost','Grocery Area'],axis = 1)

numeric_columns  = df.drop(['Video Store','Florist','Coffee Bar','Ready Food'],axis = 1).select_dtypes(include='number')
categorical_columns = df.select_dtypes(include='category')

print(f'len of numeric_columns: {len(numeric_columns.columns)}')
print(f'len of categorical_columns: {len(categorical_columns.columns)}')

df['Promotion Name'].isnull().sum()

df = df.dropna(subset=['Promotion Name'], axis=0)

pd.set_option('display.max_columns', 23)
df.head()

df.to_csv("Exploration.csv", index=False)
