==> Code/4.Model.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import category_encoders as ce
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn.impute import SimpleImputer, KNNImputer
from sklearn.preprocessing import StandardScaler, OneHotEncoder, LabelEncoder
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

df = pd.read_csv(r'Exploration.csv')

pd.set_option('display.max_columns', 23)
df.head()

df.info()

for col in df.select_dtypes(include=['object']).columns:
    df[col] = df[col].astype('category')

# Separate features and target
X = df.drop(['Cost'], axis=1)

y = df['Cost']

# Define columns by data type
numeric_cols = [col for col in X.select_dtypes(include=['float64']).columns]
categorical_cols = [col for col in X.select_dtypes(include=['category']).columns]

# Assuming you have defined X and y elsewhere

# Split your data into train and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.10, random_state=42, shuffle=True)

# Define your preprocessing steps for numeric data
numeric_transformer = Pipeline(steps=[
    ('imputer', KNNImputer(n_neighbors=2)),
    ('scaler', StandardScaler())
])

# Preprocessing for categorical data
categorical_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='most_frequent')),
    ('BinaryEncoder', ce.BinaryEncoder(return_df=True, handle_unknown='ignore'))

])

# Define the numeric and categorical column names
numeric_cols = X_train.select_dtypes(include=[np.number]).columns.tolist()
categorical_cols = X_train.select_dtypes(exclude=[np.number]).columns.tolist()

# Create the ColumnTransformer
preprocessor = ColumnTransformer(
    transformers=[
        ('num', numeric_transformer, numeric_cols),
        ('cat', categorical_transformer, categorical_cols)
    ]
)

# Create the main pipeline including preprocessing and the model
pipeline = Pipeline(steps=[
    ('preprocessor', preprocessor),
    ('model', RandomForestRegressor(random_state=42))
])

# Define the parameter grid to search
param_grid = {
    'model__n_estimators': [50, 100, 200],
    'model__max_depth': [None, 10, 20],
    'model__min_samples_split': [2, 5, 10],
    'model__min_samples_leaf': [1, 2, 4]
}

# Create RandomizedSearchCV with cross-validation
Randomized_search = RandomizedSearchCV(
    estimator=pipeline,
    param_distributions=param_grid,
    scoring='neg_mean_squared_error',
    cv=5,
    verbose=2,
    n_jobs=-1
)

# Fit the randomized search to your data
Randomized_search.fit(X_train, y_train)

# Get the best parameters and best estimator
best_params = Randomized_search.best_params_
best_rf_model = Randomized_search.best_estimator_

# Make predictions on the test set using the best_rf_model
predictions = best_rf_model.predict(X_test)


# Make predictions on the test set
y_pred = best_rf_model.predict(X_test)

# Calculate RMSE
rmse_test = mean_squared_error(y_test, y_pred, squared=False)
print(f"Root Mean Squared Error test: {rmse_test}")
y_train_pred = best_rf_model.predict(X_train)
# Calculate RMSE
rmse_train = mean_squared_error(y_train, y_train_pred, squared=False)
print(f"Root Mean Squared Error train: {rmse_train}")

import joblib

joblib.dump(best_rf_model, 'final_model.joblib')


X.columns
