==> Code/5.test.ipynb <==
import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns 
import re

df = pd.read_csv(r'C:\Users\SALAH MAHMOUD\Downloads\ieee-mansb-victoris-2\test.csv')

df.head()

df.info()

df.rename(columns = {'Unnamed: 0':'ID'}, inplace = True)

for col in df.select_dtypes(include=['object']).columns:
    df[col] = df[col].astype('category')

# Removing 'Millions' and converting to float for train_1
for col in ['Store Sales', 'Store Cost']:
    df[col] = df[col].astype(str).str.replace('Millions', '').str.replace('illions', '').str.replace('M', '').astype(float) * 1000000


# Function to extract city&country codes
def extract_city_code(place_code):
    parts = place_code.split('_')  # Split by underscore
    if len(parts) == 2:
        return parts[0]
    else:
        return None
# Function to extract city&country codes
def extract_country_code(place_code):
    parts = place_code.split('_')  # Split by underscore
    if len(parts) == 2:
        return parts[1]
    else:
        return None

# Apply the function to create a new 'City Code' column
df['City Code'] = df['Place Code'].apply(extract_city_code)
df['Country Code'] = df['Place Code'].apply(extract_country_code)

df.drop('Place Code',axis=1,inplace = True)

# Define extraction functions
def extract_status(text):
    status_pattern = r'Single|Married'
    status_match = re.search(status_pattern, text)
    return status_match.group() if status_match else None

def extract_children(text):
    children_pattern = r'\b(\w+) children\b'
    children_match = re.search(children_pattern, text)
    return children_match.group(1) if children_match else None

def extract_education(text):
    education_pattern = r'education:\s*(\w+\s*\w+)'
    education_match = re.search(education_pattern, text, re.IGNORECASE)
    return education_match.group(1) if education_match else None

def extract_working(text):
    working_pattern = r'as (\w+(?: \w+)*)'
    working_match = re.search(working_pattern, text, re.IGNORECASE)
    return working_match.group(1) if working_match else None

def extract_sex(text):
    sex_pattern = r'\b(?:Male|Female)\b'
    sex_match = re.search(sex_pattern, text, re.IGNORECASE)
    return sex_match.group() if sex_match else None


# Loop through each DataFrame
df['Status'] = df['Person Description'].apply(extract_status)
df['Number of Children'] = df['Person Description'].apply(extract_children)
df['Education'] = df['Person Description'].apply(extract_education)
df['Working Status'] = df['Person Description'].apply(extract_working)
df['Sex'] = df['Person Description'].apply(extract_sex)

df.drop('Person Description',axis=1,inplace = True)

# Extracts product type from text
def extract_product_type(text):
    pattern = r'^([\w\s]+) from' 
    match = re.search(pattern, text)
    return match.group(1).strip() if match else None

# Extracts ordered brand from text
def extract_ordered_brand(text):
    pattern = r'Ordered Brand : ([\w\s]+)' 
    match = re.search(pattern, text)
    return match.group(1).strip() if match else None

# Extracts department from text
def extract_department(text):
    pattern = r'from ([\w\s]+) department'
    match = re.search(pattern, text)
    return match.group(1).strip() if match else None

# Loop through each DataFrame column
df['Department'] = df['Customer Order'].apply(extract_department) 
df['Ordered_Brand'] = df['Customer Order'].apply(extract_ordered_brand)  
df['Product_Type'] = df['Customer Order'].apply(extract_product_type)  

df.drop('Customer Order',axis=1,inplace = True)

# make new columns from Additional Features in market
def extract_feature(text, feature_name):
    if pd.isna(text):
        return np.nan
    elif feature_name in text:
        return 1
    else:
        return 0

# Define a list of features to extract
features_to_extract = ['Video Store', 'Florist', 'Coffee Bar', 'Bar For Salad', 'Ready Food']

# Loop through the list of features and apply the extraction function
for feature in features_to_extract:
    df[feature] = df['Additional Features in market'].apply(lambda x: extract_feature(x, feature))

df.drop('Additional Features in market',axis=1,inplace = True)

# Mapping of textual values to numerical values
mapping = {
    'No': 0,
    'one': 1,
    'two': 2,
    'three': 3,
    'four': 4,
    'five': 5,
    # Add more mappings as needed
}

# Replace textual values with numerical values
df['Number of Children'] = df['Number of Children'].replace(mapping)

# Convert the column to numeric (this will handle NaN values as well)
df['Number of Children'] = pd.to_numeric(df['Number of Children'], errors='coerce')


df['Yearly Income'] = df['Yearly Income'].str.replace("K", "").str.replace("+", "").str.replace(".", "").astype('float64')


pd.set_option('display.max_columns', 28)
df.head()

# Define a function 'Income' that categorizes income values
def Income(x):
    if x <= 70:
        return 'less_70'  # If income is less than or equal to 70, categorize as 0
    elif x > 70:
        return 'greater_70'  # If income is greater than 70, categorize as 1

df['Yearly Income'] = df['Yearly Income'].apply(Income)

def Is_Recyclable(x):
    if x == 'recyclable':
        return 'yes'
    elif x == 'non recyclable':  
        return 'no'
    elif x == 'yes':
        return 'yes'
    elif x == 'no':return 'no'

df['Is Recyclable?'] = df['Is Recyclable?'].apply(Is_Recyclable)

for col in df.select_dtypes(include=['object']).columns:
    df[col] = df[col].astype('category')

pd.set_option('display.max_columns', 28)
df.head()

expected_column_order = ['ID','Promotion Name', 'Store Kind', 'Store Sales', 'Is Recyclable?',
       'Yearly Income', 'Store Area', 'Meat Area', 'Gross Weight',
       'Package Weight', 'City Code', 'Country Code', 'Status',
       'Number of Children', 'Working Status', 'Sex', 'Department',
       'Ordered_Brand', 'Product_Type', 'Video Store', 'Florist', 'Coffee Bar',
       'Ready Food']

df = df[expected_column_order]

X_test = df.drop(['ID'],axis = 1)

# Define columns by data type
numeric_cols = [col for col in X_test.select_dtypes(include=['float64']).columns]
categorical_cols = [col for col in X_test.select_dtypes(include=['category']).columns]

import joblib

model = joblib.load('final_model.joblib')

# Assuming predictions is a multi-dimensional array
predictions = model.predict(X_test)

# Create submission DataFrame
submission_df = pd.DataFrame({
    "ID": df["ID"].values,
    "Cost": predictions
})

# Save submission as CSV
submission_df.to_csv("final_test.csv", index=False)
